Move common.py file to the training platform root
