+++ tests/training_platform/test.py
@@ -23,7 +23,7 @@ with open(config_file_path, 'w') as configfile:
 
 import unittest
 
-from training_platform.server.common import *
+from training_platform.common import *
 from environments.tic_tac_toe.tic_tac_toe_engine import TicTacToeEngine
 from reinforcement_learning.agents.basic_mc_agent.basic_mc_agent import BasicAgent
 from training_platform import EnvironmentServer
@@ -31,8 +31,7 @@ from training_platform import AgentClient
 from training_platform.server.environment_server import EnvironmentNotReadyToStartError, \
                                                         AccessingUninitializedEnvServerError,\
                                                         EnvServerReinitializingError
-from training_platform.clients.agent_client import MatchMakerUninitializedError,\
-                                                   InvalidPlayer,\
+from training_platform.clients.agent_client import InvalidPlayer,\
                                                    RejoiningError
