+++ training_platform/config.ini
