+++ training_platform/server/environment_server.py
@@ -8,10 +8,10 @@ sys.path.append(ABS_PROJECT_ROOT_PATH)
 # -------------------------PROJECT-ROOT-PATH-APPENDING----------------------END#
 
 from thespian.actors import *
-from training_platform.server.common import *
+from training_platform.common import *
 from training_platform.server.service import GameManager, MatchMaker
 from training_platform.server.logger import Logger
-from training_platform.server.common import LOGGING
+from training_platform.common import LOGGING
 
 
 class EnvironmentNotReadyToStartError(Exception):
