+++ training_platform/server/logger.py
@@ -8,7 +8,7 @@ sys.path.append(ABS_PROJECT_ROOT_PATH)
 # -------------------------PROJECT-ROOT-PATH-APPENDING----------------------END#
 
 from thespian.actors import *
-from training_platform.server.common import *
+from training_platform.common import *
 
 class Logger(Actor):
     def __init__(self):
